update Settings component file path
