--- client/src/components/chat/settings.css
@@ -1,350 +1,0 @@
-body {
-    font-family: 'Arial', sans-serif;
-    margin: 0;
-    padding: 0;
-    background-color: #f4f4f4;
-}
-
-.settings-container {
-    display: flex;
-    height: 100vh;
-}
-
-.sidebar {
-    width: 250px;
-    background-color: #882A85;
-    color: white;
-    padding: 20px;
-    box-shadow: 2px 0 5px rgba(0, 0, 0, 0.1);
-}
-
-.sidebar-header {
-    display: flex;
-    align-items: center;
-    margin-bottom: 30px;
-}
-
-.sidebar-header img {
-    width: 20px;
-    height: 20px;
-    margin-right: 10px;
-    margin-top: 4px;
-}
-
-.sidebar h3 {
-    font-size: 25px;
-    margin: 0;
-}
-
-.sidebar ul {
-    padding: 0;
-}
-
-.sidebar ul li {
-    padding: 15px 15px;
-    margin: 10px 0;
-    font-size: 18px;
-    border-radius: 5px;
-    cursor: pointer;
-    transition: background-color 0.3s ease, color 0.3s ease;
-}
-
-.sidebar ul li.active {
-    background-color: #9c5199;
-    color: #fff;
-}
-
-.sidebar ul li:hover {
-    background-color: #9c5199;
-}
-
-.main-content {
-    flex: 1;
-    padding: 20px;
-    background-color: white;
-    box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.1);
-    overflow-y: auto;
-}
-
-.content h3 {
-    font-size: 24px;
-    margin-bottom: 20px;
-}
-
-.content label {
-    display: block;
-    font-size: 20px;
-    margin-bottom: 8px;
-    color: #882A85;
-}
-
-.content input[type="text"],
-.content input[type="password"],
-.content input[type="email"],
-.content select {
-    width: 100%;
-    padding: 10px;
-    font-size: 16px;
-    margin-bottom: 15px;
-    border: 1px solid #ddd;
-    border-radius: 4px;
-    box-sizing: border-box;
-}
-
-.content button {
-    background-color: #882A85;
-    color: white;
-    padding: 10px 20px;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-    transition: background-color 0.3s ease;
-}
-
-.content button:hover {
-    background-color: #9c5199;
-}
-
-.alert-box {
-    background-color: #e74c3c;
-    color: white;
-    padding: 15px;
-    border-radius: 4px;
-    margin-top: 30px;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-}
-
-.alert-box p {
-    margin: 0;
-    font-size: 16px;
-}
-
-.alert-box button {
-    background-color: #c0392b;
-    padding: 8px 16px;
-    font-size: 14px;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-    transition: background-color 0.3s ease;
-}
-
-.alert-box button:hover {
-    background-color: #e74c3c;
-}
-
-.switch {
-    position: relative;
-    display: inline-block;
-    width: 45px;
-    height: 24px;
-    margin-left: 10px;
-}
-
-.switch input {
-    display: none;
-}
-
-.slider {
-    position: absolute;
-    top: 0;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    background-color: #ccc;
-    transition: 0.4s;
-    border-radius: 24px;
-}
-
-.slider:before {
-    position: absolute;
-    content: "";
-    height: 20px;
-    width: 20px;
-    left: 4px;
-    bottom: 2px;
-    background-color: white;
-    transition: 0.4s;
-    border-radius: 50%;
-}
-
-input:checked + .slider {
-    background-color: #882A85;
-}
-
-input:checked + .slider:before {
-    transform: translateX(18px);
-}
-
-.privacy-options {
-    display: flex;
-    flex-direction: column;
-}
-
-.privacy-option {
-    margin-bottom: 20px;
-}
-
-.toggle-button {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    width: 100%;
-    padding: 10px;
-    background-color: #eeeeee;
-    border: 1px solid #ddd;
-    cursor: pointer;
-    text-align: left;
-    transition: background-color 0.3s ease;
-}
-
-.toggle-button:hover {
-    background-color: #ddd;
-}
-
-.option-content {
-    margin-top: 10px;
-    padding: 10px;
-    background-color: #f9f9f9;
-    border: 1px solid #ddd;
-    border-radius: 4px;
-}
-
-.option-content p {
-    margin: 0 0 10px;
-    font-size: 16px;
-}
-
-.option-content label {
-    display: block;
-    margin-bottom: 5px;
-}
-
-.notification-section {
-    margin-bottom: 20px;
-    padding: 20px;
-    border: 1px solid #e0e0e0;
-    border-radius: 10px;
-    background-color: #f9f9f9;
-    box-shadow: 0px 2px 10px rgba(0, 0, 0, 0.05); 
-}
-
-.notification-title {
-    font-size: 16px;
-    font-weight: 500;
-    margin-bottom: 20px;
-    color: #882A85;
-    border-bottom: 2px solid #ddd;
-    padding-bottom: 10px;
-}
-
-.notification-item {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 15px 0;
-    border-bottom: 1px solid #ddd;
-}
-
-.notification-item:last-child {
-    border-bottom: none; 
-}
-
-.notification-label {
-    font-size: 18px;
-    color: #555;
-    flex: 1; 
-    margin-right: 20px;
-}
-
-.switch {
-    position: relative;
-    display: inline-block;
-    width: 50px;
-    height: 26px;
-}
-
-.switch input {
-    display: none;
-}
-
-.slider {
-    position: absolute;
-    cursor: pointer;
-    top: 0;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    background-color: #ccc;
-    transition: 0.4s;
-    border-radius: 26px;
-}
-
-.slider:before {
-    position: absolute;
-    content: "";
-    height: 22px;
-    width: 22px;
-    left: 2px;
-    bottom: 2px;
-    background-color: white;
-    transition: 0.4s;
-    border-radius: 50%;
-}
-
-input:checked + .slider {
-    background-color: #882A85;
-}
-
-input:checked + .slider:before {
-    transform: translateX(24px);
-}
-
-
-.change-password-container {
-    margin-top: 20px;
-}
-
-.change-password-button {
-    background-color: #9c5199;
-    color: white;
-    padding: 10px 20px;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-    transition: background-color 0.3s ease;
-}
-
-.change-password-button:hover {
-    background-color: #9c5199;
-}
-
-.delete-account-button {
-    background-color: #ad2626;
-    color: white;
-    padding: 10px 20px;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-    transition: background-color 0.3s ease;
-    margin-top: 20px;
-}
-
-.delete-account-button:hover {
-    background-color: #c0392b;
-}
-
-.blocked-contacts {
-    margin-top: 15px;
-    padding: 10px;
-    border: 1px solid #ddd;
-    background-color: #f9f9f9;
-    border-radius: 4px;
-}
-
-.blocked-contacts p {
-    font-size: 16px;
-    margin: 8px 0;
-}
